+++ src/commands/exit/exit.js
