+++ src/commands/exit/index.js
