+++ src/commands/filesSystem/add.js
@@ -1,7 +1,7 @@
 import { writeFile } from 'fs/promises';
 import { join } from "path";
 
-import { getCurrentDirectory } from "../../../utils/index.js";
+import { getCurrentDirectory } from "../../utils/index.js";
 
 export const add = async ({ directory, argument }) => {
 	if(!argument) {
