+++ src/commands/filesSystem/rm.js
@@ -1,6 +1,6 @@
 import { unlink } from 'fs/promises';
 
-import {createAbsolutePath, getCurrentDirectory} from "../../../utils/index.js";
+import {createAbsolutePath, getCurrentDirectory} from "../../utils/index.js";
 
 export const rm = async ({ directory, argument }) => {
 	if(!argument) {
