+++ src/commands/filesSystem/rn.js
@@ -1,6 +1,6 @@
 import { rename } from 'fs/promises';
 
-import {checkArguments, getCurrentDirectory} from "../../../utils/index.js";
+import {checkArguments, getCurrentDirectory} from "../../utils/index.js";
 
 
 export const rn = async ({ directory, argument, additionalArgument }) => {
