+++ src/commands/hash/index.js
