+++ src/commands/moving/cd.js
@@ -1,7 +1,7 @@
 import { access } from 'fs/promises';
 import { resolve } from "path";
 
-import {createAbsolutePath, getCurrentDirectory, setCurrentDirectory} from "../../../utils/index.js";
+import {createAbsolutePath, getCurrentDirectory, setCurrentDirectory} from "../../utils/index.js";
 
 export const cd = async ({ directory, argument }) => {
 	if(!argument) {
