+++ src/commands/moving/index.js
