+++ src/commands/moving/up.js
@@ -1,6 +1,6 @@
 import { join } from "path";
 
-import { getCurrentDirectory, setCurrentDirectory } from "../../../utils/index.js";
+import { getCurrentDirectory, setCurrentDirectory } from "../../utils/index.js";
 
 export const up = async({ directory }) => {
 	const newDirectory = join(directory, "..")
