+++ src/commands/zip/index.js
